+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.6.3] - 2021-03-26
+
+### Change
+
+* Only final cert in chain used to verify when signing PKINIT requests, /verifycerts flag added for full verification
+
+### Fixed
+
+* Replaced /certificate info for help and README
+
+
 ## [1.6.2] - 2021-03-12
 
 ### Added